requirements.txt
tensorflow==2.16.2
streamlit==1.50.0
numpy==1.26.4
pillow==11.3.0
scikit-learn==1.7.2